Add files via upload
fixed loops

+++ sc/cleanup.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
 # Clean up wget
+if test -f ./chrome.deb 
+then
 rm chrome.deb
+fi
 rm code.deb 
 rm fastfetch.deb
 echo ""
